final commit

+++ subscription/src/main/java/com/subscriptionTable/subscription/dao/SubscriptionDaoInterface.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.subscriptionTable.subscription.entity.SUBSCRIPTION_TABLE;
 
 @Repository
-public interface SubscriptionDaoInterface extends JpaRepository<SUBSCRIPTION_TABLE, Integer>{
+public interface SubscriptionDaoInterface extends JpaRepository<SUBSCRIPTION_TABLE, Long>{
 
 }
